9. setup basic configuration for the frontend

+++ backend/controller/product.controller.js
@@ -21,13 +21,17 @@ export const createProduct = async (req, res) => {
 export const deleteProduct = async (req, res) => {
   const { id } = req.params; // get the id from the url
 
+  if (!mongoose.Types.ObjectId.isValid(id)) {
+    return res.status(404).json({ success: false, message: 'Invalid product ID' }); // check if the id is valid
+  }
+
   try {
     await Product.findByIdAndDelete(id); // delete the product from the database
     console.log(`Product id: ${id} successfully deleted!`);
     res.status(200).json({ success: true, message: 'Product deleted successfully' });
   } catch (error) {
     console.log('Error deleting product:', error.message);
-    res.status(404).json({ success: false, message: 'Product not found' });
+    res.status(500).json({ success: false, message: 'Server error' });
   }
 };
